Show dataset events above task/run details in grid view (#37603)
* Show dataset events above task/run details in grid view

* Add extra to dataset event tables
