+++ airflow/www/static/js/dag/details/dagRun/DatasetTriggerEvents.tsx
@@ -17,9 +17,10 @@
  * under the License.
  */
 import React, { useMemo } from "react";
-import { Box, Heading, Text } from "@chakra-ui/react";
+import { Box, Text } from "@chakra-ui/react";
 
 import {
+  CodeCell,
   DatasetLink,
   Table,
   TaskInstanceLink,
@@ -55,6 +56,12 @@ const DatasetTriggerEvents = ({ runId }: Props) => {
         accessor: "timestamp",
         Cell: TimeCell,
       },
+      {
+        Header: "Extra",
+        accessor: "extra",
+        Cell: CodeCell,
+        disableSortBy: true,
+      },
     ],
     []
   );
@@ -63,7 +70,9 @@ const DatasetTriggerEvents = ({ runId }: Props) => {
 
   return (
     <Box mt={3} flexGrow={1}>
-      <Heading size="md">Dataset Events</Heading>
+      <Text as="strong" mb={3}>
+        Dataset Events
+      </Text>
       <Text>Dataset updates that triggered this DAG run.</Text>
       <Table data={data} columns={columns} isLoading={isLoading} />
     </Box>
